新增动画；替换关卡；新增模型

+++ PigRun/Assets/PIgGame/Scripts/LevelData.cs
@@ -7,12 +7,12 @@ public class LevelData
 {
     public Size size;                      // 地图世界大小
     public float time;                      // 倒计时
-    public int tap;                          // 点击次数
+    public float tap;                          // 点击次数
     public CameraPosition cameraPos;         // 相机位置
     public CameraAngle cameraAngle;          // 相机角度
     public List<PigData> pigGroup;           // 猪的列表
     public List<object> obstacleGroup;       // 障碍物（空数组）
-    public int roadSpeed;                     // 道路速度
+    public float roadSpeed;                     // 道路速度
     public bool is2Dir;                        // 是否双方向
     public bool isEatAnim;                     // 是否有吃动画
 }
@@ -40,9 +40,9 @@ public class CameraAngle
 public class PigData
 {
     public Position position;
-    public int angle;      // 旋转角度（0/90/180/270）
-    public int type;        // 猪的类型，对应不同预制体
-    public int boomTime;    // 爆炸时间（可能未使用）
+    public float angle;      // 旋转角度（0/90/180/270）
+    public float type;        // 猪的类型，对应不同预制体
+    public float boomTime;    // 爆炸时间（可能未使用）
 }
 
 [System.Serializable]
